requirements.txt
# --------- pytorch --------- #
lightning==2.0.9.post0
pytorch-lightning==2.0.9.post0
torch==2.0.1
torchvision==0.15.2

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
# hydra-ax-sweeper
# hydra-ray-launcher
# hydra-submitit-launcher

# --------- loggers --------- #
wandb==0.18.5
# neptune-client
# mlflow
# comet-ml
# torch_tb_profiler

# --------- linters --------- #
pre-commit==4.0.1      # hooks for applying linters on commit

# --------- others --------- #
# python-dotenv   # loading env variables from .env file
rich==13.9.3
scikit-learn==1.5.2
seaborn==0.13.2
rasterio==1.4.1
fiona==1.10.1
pyarrow==13.0.0
geopandas==1.0.1
numpy==1.25.2
scipy==1.14.1
scikit-image